// frontend/src/api/generated/model/organizationMemberResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CloudOps API
 * Infrastructure resource management, monitoring and asynchronous tasks. Timestamps use UTC ISO-8601. Error responses share ApiError; errors contains field violations.
 * OpenAPI spec version: 0.0.1-SNAPSHOT
 */
import type { OrganizationMemberResponseRole } from "./organizationMemberResponseRole";

export interface OrganizationMemberResponse {
  id?: number;
  organizationId?: number;
  userId?: number;
  role?: OrganizationMemberResponseRole;
  createdAt?: string;
  updatedAt?: string;
}